Abins2D: fix DW application when no autoconvolution
This was an oversight leading to a "referenced before assignment"
error. The issue was already fixed in the Abins1D code, but in this
duplicated region the fix was overlooked. Most manual testing used
autoconvolution so it wasn't noticed for a while.

Anyway, this duplication needs to be cut down somehow.