Add ReflectometryWorkflowBase tests and minor refactoring
- Add a test file for ReflectometryWorkflowBase2. This is an abstract class but has a lot of untested functionality that is not tested in the derived classes so I've added the file so we can start adding tests for these functions. Due to it being abstract with protected members I've created a stub class derived from it which is a friend of the tests.
- I've started by adding tests for findProcessingInstructions due to upcoming work planning to change this.
- I've done some minor refactoring to move duplicated code that declares the AnalysisMode property into the base class because this is required by the tests.

Note that I have not been able to add tests for the invalid cases where the function throws if properties are missing from the parameters file. This is because of a quirk in Mantid where the previously-loaded parameters file for the same instrument is loaded when you call LoadInstrument, i.e. it then starts with populated parameters for the ones I am trying to test that are empty, and even when I re-load the correct parameters file (where they are missing) they remain populated. It is beyond the scope of this commit to fix this at this point.

Re #33640