Fix CMake option call to set expected initial values
The second argument to option in CMake is a doc string the initial value is defaulted to OFF
if it is not provided. We were accidentally setting everything to OFF so parts of the configuration
were not behaving as expected. This fixes the calls so we now skip elements of the Eigen
config that we don't need.