Log warnings when using default settings for live data
I've added warnings both in the RunsPresenter when requesting the defaults, as well as in createAlgorithmRuntimeProps when it searches for the wildcard row. The former allows us to warn specifically about live data, and the latter to say specifically which defaults are used. Although this is slight duplication I think it is better to be over-informative here because it is very confusing if unexpected defaults get used. The latter function is not used anywhere else at the moment but the warnings will still make sense if it is in the future.

Re #33649