Fix spurious warnings when running reduction
createAlgorithmRuntimeProps was calling the overloaded version of the same function that gets the wildcard row first even though it will be overridden by the one specific to the row. Since the function now gives a warning if a wildcard row is used this was giving a spurious warning. I've done some refactoring to have a single function do all of the work, which now takes an optional row.

I've moved the warnings from the LookupTable into RowProcessingAlgorithm so they are all in one place and can be managed more consistently.

Re #33649