Drop duplicated norm/trans monitor selection
Drops duplicated selector for normalisation and transmission monitors.
We will instead specify the monitor name (e.g. M1) then go and lookup
this in the normalisation block.

Additionally adds tests and some hard-coded assumptions that
instrument.configuration always uses spectra numbers. Now both will
lookup the spec num from the relevant block.

Updates the transmission test to better reflect we can have a different
normalisation. In a future commit we will add support for this and the
relevant tests, but at the moment add a test to ensure nobody hardcodes
this assumption.

Re #33659