Fix failing python api start on debug ubuntu
Attempting to initialize the framework manager from the python API (like
is done in the reflectometry Encoder and Decoder tests) was causing a
segfault due to some undefinfed behaviour as a result of certain
variables dropping out of scope between the python and C++ layer.
The imports have now been removed, instead adding an algorithm to the
factory for testing and removing it afterwards. This algorithm is used
for getting the properties from only.