Add lookupIndex info to Run rows
In order for the runs table to track which lookup rows are being used,
each run row needs to track which of the run rows it is going to be
processed by. A value of boost::none means that no lookup row was found
and the algorithm defaults will be used.

RE #33624