Add a fallback stream operator for boost::optional
GoogleTest fixed a bug in discovering operator<< by
ADL, including finding an operator for a base type.
boost::optional provides a default operator<< that
asserts a separate file, optional_io.hpp, should be
included to output the type. Including the file requires
the value type of the optional to  have a defined
stream operator.
We simply provide a basic boost::optional printer
that is sufficient for tests rather than polluting
the main code with operators that are not
useful.