Tighten up validation on S(Q,w) for inelastic calculations
If S(Q,w) only contained values for positive w (energy loss) then some calculations were
failing in SampleKW because there were no available w values in S(Q,w) that allowed a
path to be completed if the neutron's energy had already dropped close to zero
Fix by requiring S(Q,w) to contain some negative w (energy gain) values and add unit test
to cover this. Annoyingly one of the calculation result tests used an S(Q,w) with only positive
values so I've also updated that so that the spike in the test S(Q,w) is at -ve rt +ve w

Also spotted that the importance sampling for inelastic assumed the S(Q,w) had constant
bin width in w. This is the case in all the test inputs I've used but might not always
be so have multiplied by the w bin width in the code when calculating the probability
distribution for Q, w

Also tidy up use of braces in ValidateInputs on single line if statements and append to
issues list rather than overwrite in validations on StructureFactorWorkspace