Modernise SANS2D Tests
Replaces the SANS2D tests from pre-2015 with more modern versions.
This is because the workflow does not represent the fact that all data
is now collected with monitor mode. This also means we can't map 1:1
from legacy files to new files as the new command is not the same.

Re #33659