Widen tolerance for unit test
The changes in an earlier commit to switch the elastic input units
from wavelength to momentum had shifted the actual value so that it
was very close to the 1e-05 tolerance. This was due to the bin order
being switched and the test bin became [2] instead of [0]. This in
turn meant the random numbers used for bin [2] were different.

The last commit on this change to support containers moved it by a tiny
amount and it has dropped over the 1e-05 threshold. Note - the reference
value is still correctly centred it's just the spread is bigger than
originally thought