Fix regression of bug #33104 reveresed in PR #33639
Bug relates to support for dynamic rebinning when MDhisto workspace has been altered (e.g. scaled).
It was initially fixed in #33362 which introduced a function to determine whether the orignal
MDE workspace should be rebinned (it shouldn't if the MDHisto has been altered). In PR #33639
the function that checks whether a ws can be dynamically rebinned was moved and it no longer
checked whether the MDHisto had been altered (but the function to do so remained in the model).
This commit moves the function to check if the MDHisto had been altered to the WorkspaceInfo model
and fixes the regression.