Fix a bug where unsupported toml files stop GUI opening
The SANS GUI remembers previously-opened user files. If you had previously opened a v0 toml file in a version of Mantid that supported v0, the SANS GUI will attempt to reload it. This will now fail because v0 is no longer supported. This is expected; however, there was a bug where this was causing an unhandled exception which was stopping the GUI opening at all. This commit wraps the exception in the same way as other expected exceptions meaning that it is now handled correctly.

Re #33718