Release GIL before posting ADS notifications
This commit fixes a bug where Mantid could hang due to a deadlock between the GIL and a lock set when posting notifications about ADS updates. If an ADS function such as addOrRemove is called from a python algorithm (which has the GIL locked) then the poco notification attempts to take a lock. However, this could already be locked by another poco notification, and it could be possible for the other notification to require the GIL in which case we will end in deadlock.  The solution is to ensure we release the GIL in any ADS functions that are exported to python that make modifications.

Re #33664