Abins: Use Euphonic Castep parser where available
The inbuilt Castep parser makes the unsafe assumption that
Gamma-points with LO-TO splitting, which use a longer header line to
contain extra information, will always appear first. If this point
appears later in the data, earlier k-points are split.

This assumption is quite deeply baked into the logic of the parser,
requiring refactoring to address. As we intend in any case to
increasingly depend on Euphonic, which contains a .phonon file parser
/without/ this defect, this work would be redundant.

Here we
- detect the unsafe behaviour in the inbuilt parser and raise an error
- use Euphonic if possible

Not all Mantid builds include Euphonic yet, but when we move to
Conda-based packaging for all builds the legacy implementation can be
removed.

Abins is structured around the idea that each format has a class
implementing the loader, so I have created a wrapper class around the
two implementations. This will make it very easy to remove the old
implementation in future. For now, the use_euphonic flag can be used
in testing to force one implementation or the other.