Fix bug in the python export of addExperimentInfo
We need to use ExtractSharedPtr to extract the actual shared ptr from the python object; otherwise it will create a separate shared_ptr. This was causing a crash when destroying the object, although this was only reproducible in certain niche circumstances (on conda, when using the new ADS clear function that releases the GIL, and only when done in the tearDown function of the test). This commit also adds a test for this case.

Co-authored-by: Martyn Gigg <martyn.gigg@stfc.ac.uk>

Re #33664