Remove spurious additional factor of 2
Remove from documentation of importance samplings
The factor of 2 isn't in the code and shouldn't be there. I think this
was put in to support having a factor of 2 in the denominator (2k^2)
but you don't need to force this. There's a factor of 2 in the denominator
naturally once the integration over phi is done