Fix misleading cell state checking tests
Cell comparison does not take the colour into account. This is for ease
elsewhere. This adds a custom matcher to ensure that the colours are
also checked, as they represent the state. Also moves the setting of the
group state after the setting of the row state. Rows update their
parents when their state is updated so need to have their states set
last.

RE #33474