Reset memory heavy variables inside SmoothNeighbours
Algorithm objects are not immediately deleted to give objects
subscribed to finish notifications a chance to run and
still see a live object. SmoothNeighbours held on to
its input workspace so kept this workspace alive for
as long as the algorithm lives. For large workspaces
this causes a delay in freeing up memory
(possibly 10s of GB) and might cause operations to
fail if the process is at the limit of available
memory.

This is admittedly not an ideal solution but solves
an immediate problem for WISH users.