Fix crash when group workspace is validated
When validating inputs the conversion of the index property to an index
set was causing unhandled exceptions that could crash mantid if run and
would interfere with the doc tests if not. This adds a catch for that
behaviour and also warns the user that they may not use a group
workspace for this algorithm.

RE #33627