Abins Castep import: update systemtest
This is an unfortunate case: the results are expected to depend on the
installed libraries, because the inbuilt and Euphonic loaders handle
some sample files differently. (This is due to the ambiguous weighting
of LO-TO-split Gamma point data in .phonon files.)

As the systemtest cannot change the low-level arguments to force one
backend or the other, and could plausibly be run on either system,
we provide test data for both cases.

When Conda migration is complete, we should be able to remove the
legacy loader and simplify things.