Abins Castep/Euphonic import: simple mocking test
I'm not 100% happy with this as the "except" could catch a lot of
different things. Better would perhaps be to mock a return value
for QpointPhononModes.from_castep, then we can thoroughly check
the code in Abins and _only_ in Abins.

But then we would be entirely reliant on the systemtest to catch
upstream changes. Perhaps that's good: upstream changes are then
caught in only one or two places, so much easier to respond to with
updated data.

Anyway, to improve this further we really need better test coverage of
the EuphonicLoader and its data reformatting methods. Which might be a
separate bit of work, and coupled to caching?