Fix deadlock with algorithm manager python exports
The AlgorithmManager has a mutex in some functions which are exposed to Python. It is possible for Python code that is holding the GIL to call these and request the mutex. However it is possible the mutex may already be held by another thread running C++ code. If the other thread is using the AlgorithmManager to try to create a Python algorithm then it too will require the GIL. This can result in deadlock.

This commit fixes the problem by ensuring that the AlgorithmManager exports always release the GIL before calling into C++ code that requires the mutex.

Re #33894