BUGFIX: DensityOfStates .phonon import
This resolves a nasty bug in the CASTEP .phonon file importer used by
the SimulatedDensityOfStates Algorithm and
Indirect/Simulation/DensityOfStates interface.

There is some ambiguity in CASTEP .phonon files when data includes
LO-TO splitting. This introduces a discontinuity at the "Gamma
point" (q = [0, 0, 0]) which is sometimes addresed by computing values
in several directions. A CASTEP .phonon file from such a calculation
will include entries for each of these, and they will each be given
the full "weight" value of Gamma. If those are naively used, this will
lead to over-weighting of the Gamma-point compared to other q-points.

The _intended_ strategy of this .phonon reader is to only take the
first Gamma point provided and reject others. It does this by

- allowing the first q-point regardless of its q-vector, assuming that
  in such LO-TO meshes any such Gamma-point will be first in the list.

- determining if any other q-point is Gamma by checking
  `sum(q_vector) == 0`, and weighting to zero if so.

The first assumption will hold in typical user
scenarios (i.e. Monkhorst-Pack mesh generated within CASTEP) but is
not actually a safe assumption. See
https://github.com/mantidproject/mantid/pull/33761 where a similar
assumption can lead to a much worse failure case.

The second assumption, that Gamma-points can by identified with
`sum(q_vector) == 0`, is deeply flawed.

- It fails to identify (1, 1, 1) as also being Gamma due to
periodicity
- It incorrectly identifies e.g. (-0.25, 0, 0.25) as being Gamma.

In practice this leads to a small set of q-points being discarded from
the calculation arbitrarily. The worst-case scenario would be a
calculation that follows e.g. the (0, 0, 0) -> (0.5, 0.5, 0) high
symmetry line, where all those points would be removed. (This would
not be a good DOS sampling even if implemented correctly, but maybe
the user has their own reasons and knows what they are doing. It would
be a valid file input.)

Although the qualitative damage is small in practice, this should be
treated quite seriously as the Phonon DOS may well be used as an input
to further calculations that are sensitive to this error.