Revert change that exposes lifetime issues with shared_ptrs in python
As per the comment added in this change, I am temporarily commenting out the change to release the GIL in AlgorithmManager::clear. This exposes a bug that is unrelated to this PR and is beyond the scope of this PR to fix. The deadlock problem is very unlikely to be hit via this specific function anyway so it is low priority to get the GIL fix in here.

Re #33894