Fix deadlock with GIL and logger
Ensure we release the GIL in the logger python exports. This avoids a possibility of deadlock where a python thread can have the GIL locked but get stuck waiting on a Poco mutex to issue a log message, which another thread has locked but is waiting on the GIL in order to update the GUI with the python log message.

Re #33938