Moving 6.4 release notes to manual
As part of release for 6.4 the process of dealing with the release notes moves from automatic to manual. This involves copying release notes
to what will become the final set of .rst files and moiving the seperate release notes to new Used folders.

Minor tidying of the notes has been done to prevent Sphinx Warnings on the 6.4 release notes. The conf.py file has also been updated to prevent toctree errors for the new Used folders.