Fix unhandled exception if user file not found
This commit fixes a bug where errors are not being handled correctly if a user file entered in the table in the SANS GUI cannot be found. Strangely, exceptions are caught and handled but normal errors are not dealt with, and there is an assumption that the "states" object contains valid values, which it does not, so when the code attempts to access it it causes a "tuple index out of range" error.

I've added some code to check for errors and handle these in the same way as exceptions. It might be that the current exception handling is no longer required but this needs further investigation and is beyond the scope of this PR.

I've also improved the error message to use the entered user file name rather than the full path, because the latter is empty if not found so the error is not useful.

Re #33962