Change default scaling in parameters files
This commit adds a parameter to the INTER and OFFSPEC parameters files to change the default setting for which workspace gets scaled when stitching transmission runs. The preferred option for most experiments is to scale the left-hand-side workspace, so we set the scale-RHS workspace option to false. Note that this only affects the GUI and does not change the default that is used by the algorithm, so this should be a non-breaking change.