Remove fill adjustment tests
In the process of trying to fix an issue and removing a deprecated
method, there doesn't seem to be a neat way to allow the waterfall plots
to keep their order or their fill when they are re-plotted.

As such, this fucntionality has been removed, the fill is cleared when
the lines are re-plotted.

This is the closest I can stay to the original functionality with the
warnings being removed. I expect that it could be possible but a *very*
large amount of the waterfall plotting code would likely need to be
re-written.

RE #33913