Improve error message for invalid user functions
Provide specific error message for blank user function
Improve error message for invalid user function by reporting back
the formula that's failed and the x value that was being evaluated
Also handle situation in plot better when the guess fails
to update on addition of a new function. Leave the guess_all_line
property populated so you can remove the guess.
This aligns this piece of code with other calls to
_plot_guess_workspace in the same unit where the line return value
is checked to see if it's None