Fix a bug with project recovery incorrect checkpoint
Project recovery uses a count of the current number of workbench processes versus the number of checkpoints to decide whether a crash happened. However, since moving to multiprocessing, we now have two pyton processes for each workbench instance on linux (one is the parent process for the error reporter). This messes up the calculation.

This commit fixes the problem by ignoring parent processes in the count of workbench instances.

Re #33903