Fix waterfall plot reverse order shuffle
The waterfall plot's Reverse Line Order button changed when we upgraded
to mpl v3.5. For some reason the double assignment that happens in the
reverse method

lines[i], lines[j] = lines[j], lines[i]

has stopped working. As a result, we're having to do the switch
ourselves. A lot of the waterfall functionality is under scope to be
re-written, what with accessing ax.lines being deprecated, so I don't
feel too bad about including this change, even if it is a bit ugly.

RE #34006