Convert the QSetting property to a boolean
Refs #33900. When opening multiple Mantids in quick succession, there
is sometimes a race condition between the syncing of the QSettings,
and the 'fixup' of the booleans values such as changing 'true' to
True. The syncing can be triggered by the event loop, and if this is
triggered at the correct time, the converted booleans can be
overwritten back to their original string form for a specific mantid
instance.