Relabel assets as they are uploaded to GitHub
The default label is the filename but as the filenames
are similar it can be hard to quickly pick out the correct
link to use for your platform.
We instead label each asset with a fixed string that clearly
shows the OS but also has the added bonus that existing assets
are replaced rather than added to, which is the behaviour we
desire.