Use patch version rather than date for release packages
An optional argument has been added to the update-conda-recipes script to tell it when to use a release-style version number, rather than a nightly/unstable version number. The argument is passed if no package suffix is given. Logic has been added to the jenkins nightly pipeline to implement this too.