Remove assumption that workspace exists in ADS
When masking bins, remove an assumption that the workspace the instrument viewer is showing exists in the ADS. This may not be the case going forward when instrument viewer is embedded in other widgets. There was a bug described in issue 34094 where the output workspace pointer was being set without a name, because it is assumed to have one by the algorithm. However, this can be worked around by setting a temporary name value instead.

Re #34094