Define a DebugWithRelRuntime config in MSVC
Conda only provides libraries built against
the release variant of the Visual Studio
runtime, rendering the standard Debug config
unusable.
We now define a new configuration that has
all optimizations disabled but still links
to the release version of the msvc runtime.

We also remove MinSizeRelease as a config
option that is never used.