Add functions to run refl reduction algorithm
Existing functions run the load-and-process algorithm but we also require to just run the reduction part of that process, via the algorithm ReflectometryReductionOneAuto rather than ReflectometryISISLoadAndProcess. Since the latter wraps the former, both algorithms share mostly the same properties. This commit therefore utilises most of the existing functionality to set properties on an algorithm and just adds the required functions to specify the differences. I've put the new functions in a new namespace "Reduction" rather than the existing "RowProcessing". The refactored tests mean we can easily test that both code paths produce the same results for the same inputs.

Re #34106