Add callback to update reduced workspace
Add a new callback to update the PreviewRow when the algorithm has completed (we just pass the function pointer into the configured algorithm and the existing mechanism takes care of calling it).

Set a temporary theta value hard-coded so that we can perform a reduction (if it is zero the reduction will fail).

Re #34106