Add --single-process command-line arg to workbench
This will suppress the secondary process used to run
workbench.
Crashes/segfaults of workbench itself will not be caught
while running in this mode.
It will mostly be used by debuggers so that developers get
the expected experience that starting a debug run shows a
single process where breakpoints can be attached.