Rearrange Jenkins build pipeline
mantiddocs is now only built on Linux
so must be built before any standalone
packaging is run. Conda packaging is
now run in parallel with development
environment testing to reduce build
time of the overall pipeline.

To cut down on duplicated code we introduce
a build matrix parametrized over the platform.
Any differences in platform are encapsulated in
the utility functions.