Link preview presenter and batch presenter
This commit adds a pointer to the batch presenter from the preview presenter via a subscriber interface.

Notifications have been added from the batch to the preview presenter when processing is started/paused so that the Apply button can be disabled, which is necessary to disable changing experiment settings during processing.

This is part of work in progress, and in future PRs this mechanism will be used to send a notification to the preview presenter when the Apply button is clicked.

Re #34240