Introduce QTableView item delegate to control text elision
Upgrading to Qt 5.12 caused float values in the data
view to be elided straight after the decimal point,
regardless if their was enough column width to display
more digits.
A delegate is now used to customize the display of string
values in the matrix data view such that only the last 3
characters in the cell are elided if there are more digits
to show.