Clear second transmission run if only the first is overridden
Fix a problem where the second transmission run could not be cleared if only the first is required to be set. Previously, if both were set as defaults on the Experiment Settings tab, then leaving the second clear on the Runs tab would mean the default is always used. This commit changes it so that if the first is overridden on the Runs tab, then the second is always updated from the Runs tab as well, even if it is empty.

Note that there is no way to clear the first transmission run, but this is not a problem because the first transmission run is mandatory so sould always have a value anyway in real usage