Re-start live data monitor on finish/error
The live data monitor frequently stops with an error when the data size changes or if there is a genuine error etc. Users would like it to automatically re-start in these cases. I think I have also seen cases where the algorithm "finishes" when the experiment is not finished so again users would like it to continue running. This commit automatically restarts the monitor if it has finished or throws an error.

Re #34132