Fix a bug with delayed start of live data in refl GUI
This was caused by running StartLiveData as a child algorithm in the GUI thread; for some reason this caused the live data chunk not to be processed until the second chunk of live data is received, i.e. after 20 seconds if the update interval is 20 seconds. This is annoying and also confusing for users because they see the raw data as the processed output until it updates the second time, and then they see the processed data.