Revert "Add debug code to allow developers to test live data"
This reverts commit 0d6f4b45a4f922f80fd1c5cda878fc639c738c65.

This debug code is not reliable (it relies on the config change outside of the refl gui not being overwritten by the gui before the developer starts the monitor but this is not always the case). Seeing as this is not reliable I think it is too confusing to leave in so is better to remove it. We should add proper support for testing the gui with live data in the future.