Abins DMOL3 import: Remove mass weighting
Mass weighting was originally included in this update to the parser as
it was already there. However, it's not entirely clear that is should
be.

Removing the mass weighting seems to give much better agreement with
b_traces from CRYSTAL and CASTEP calculations, so we shall assume it
is the correct behaviour.