Handle D16 old scan data differently
D16 old scan data is considered a scan data, but contrary to the new
version, it has been tested with the reduction working on histogram data
with bins, not points. This alters the results, so for this kind of data
and only this kind of data, we still use bin edges.