Only add one region if ROI button clicked again
Extra regions of 0-1 would keep being added each time one of the ROI
buttons was pressed. This adds a check so that if the user was in draw
mode, and an ROI button is clicked, then the previous region that they
were drawing is removed before adding the new one.

RE #34602