Allow live data to succeed with transmission runs
As per the comment, this is a bit of a hack but providing a dummy angle
for the live data reduction for transmission runs allows the reduction
to succeed and show a plot with similar scale to a normal run. This is
valuable when running live data because we don't want it to stop running
by encountering an error, and we don't want to delete the workspace and
lose the plot, which users will want to see updated with valid data
again during the next real run.

A better solution would be to add fully typed error handling in the
reflectometry GUI and clear the plot from there when we receive a
zero-theta error from the live data monitor. However this is quite a big
job and will require getting a custom error type from python to C++, and
adding a timer to the reflectometry RunsPresenter so that it re-tries
after a period of time. We also need to ensure that clearing the data on
the open plot does not cause errors.

Re #34574