Fix crash caused by missing method on sliceviewer base presenter
A new method was added to the SliceViewer presenter class but not to the
base class, so the other derived type, RegionSelector, was missing this
method. This caused a crash when attempting to load data into the
RegionSelector, because the image info widget was attempting to call the
new method.

Re #34700