Update logic to read input ws from ADS if present
The logic to read the input ws from ADS if present was causing
some problems on POLARIS because the focus action applies some
changes to the workspace in situ without changing the name (eg
normalises by current, subtracts empty). So a second run of the focus
action applying some reduction steps a second time on an already
partially reduced input file
Modify the logic to create a copy of the input ws so the speed up
from looking in the ADS is retained while having a clean input ws
at the start of each call to focus
Have added a switch to not copy the workspace in case some instruments
have v large workspaces that they don't want to copy because of memory
issues. This isn't plumbed into any of the instruments yet though