Modify check on whether source and sample present
Previous check involved casting source and sample to boolean. If either
object implemented __len__ then (as is case for ICompAssembly) then there
was a chance a valid source or sample ptr could evaluate the false if __len__=0

I think the source and sample in a workspace that was loaded from a file are
often of type ICompAssembly due to the code in InstrumentBuilder::addComponent.
For most cases this won't have any children meaning __len__ is zero