Enable compilation using Ninja in Visual Studio
Visual Studio requires the schema to be v2 (even though the rest of the content
is valid in the v1 preset json schema). v2 is only available in CMake 3.20 so
move cmakeMinimumRequired on in here also. My interpretation of cmakeMinimumRequired
is that it's the minimum version required to make sense of the presets json file
so not syncing this up with cmake_minimum_required in the main CMakeLists.txt

Add cmakeExecutable setting to ninja preset. This stops VS using it's internal
CMake executable and instead causes VS to look for a cmake.exe on the path - which
will be the CMake version in Conda if VS has been started up from an active conda
environment

Also rename the win-msvc preset to be win-vs since both the windows presets
use MSVC as the underlying compiler