Support override in IDF for bank positions in side by side view
Read the override bank centre coordinates from the IDF into the
Component class. Decided to this rather than base the solution
on either the GridDetector and RectangularDetector class so that
the solution will be more general and will match up better with
the types of flat detector that are supported in the side by
side view

Decided not to alter ComponentInfo in the end and instead just access
the bank centre by getting hold of the Component object itself

Move logic from AddFlatBankOfDetectors function in PanelsSurface.cpp
into the ProcessUnstructured function. This mirrors the approach for
the other types of flat detector bank (tube\structured\grid)

Remove some unnecessary references to RectangularDetector.h
- So I can see how widely it's used

Add additional rotation to process Structured so that all banks parallel
Fix minor issue in init() so that if banks don't span the origin, the origin
isn't included in m_viewRect