Allow exception handling for async fitting
An extra Keyboard interrupt was thrown in the fitting thread as well as trying
to cancel the algorithm. This may make sense for an async workflow such as a script that
may perform some processing outside of algorithms. In the case of the EngDiff fitting tab,
the Cancel button performs well, using only alg.cancel() on abort and not the extra interrupt.
Prevented any error message containing KeyBoardInterrupt (cancelled by user) from being logged
in the local _on_worker_error() method.
Also fixed a tiny bug in asynchronous where "msg" can be an empty tuple