Replace load widget enable control with whole tab
Basically anything that re-runs the reduction has the potential to cause
a crash, it's just a bit harder to do than just spamming on yhe load
button. Insead of worrying about all of the different buttons that can
cause problems individually, let's just disable all the widgets on the
tab until the reduction is complete. That should essentially erase the
danger of segfaults caused by racing algorithms that have been started
as it seems to be the clearing of validators that was causing seg-faults
all along.

RE #34745