Fix copying file backend if HDF5 has locked the file
Version 1.12 of HDF5 in Conda configured file locking
on access for the first time. If a file is open on Windows
it has an exclusive lock and a read handle cannot be
obtained to copy the file. The workaround is to close
the file, copy then reopen in the same mode.