Specify major.minor in conda_build_config
Conda-forge specifies either major.minor
or a full major.minor.patch pin.
Update to match this convention.

Also add further pins to meta yaml
to ensure consistency. Swapped qt
for pyqt as we deploy a Python package
and Qt will track this. We also
add qt-webengine for the C++ libraries.