Replace bool text  strings with bool value strings
This needs some explaining.
Since the Refl GUI is in C++, it can make calls to the C++ algorithms
fairly directly. This means that when we give it string values like
True and False they don't seem to get converted to bools properly. This
creates confusing behaviour when the user wants to be able to write in
their properties like they would in the python scripts. Instead, we
convert the bool strings to 0 and 1 strings that the algorithm is happy
with.

This is the least disruptive way I can see of solving the problem
without interfering with the algorithm or the generalised  string
processing.

RE #34844