Create Separate function and add tests
If there is an oldTie for the parameter, which was a tie to function
(e.g. f0.Height=f1.Height) then the oldTieExpression is returned.
This allows an edited Tie to function text box to revert upon an invalid entry see pr #34817
Note this function does not get an oldTie if it was custom (e.g. f0.Height=2.0)
This new function also allows fixing a parameter, when oldTie is nullptr.