Swap QGLWidget for modern QOpenGLWidget
QGLWidget has not been formally deprecated but
there are several issues with it and the docs
mark it has obselete.
On macOS it also seems impossible to correct
deal with moving an OpenGL window between
monitors with different pixel densities, while
the new widget seems to deal with it seamlessly.
We have removed almost all glViewport calls as
the base widget now handles this for us.