Set the viewport & window for the QPainter on QOpenGLWidget
By default the viewport/window fill the entire device pixel
space. For high-density pixel displays this gives
increased possibility for drawing at a finer level.
Our 2D shape drawing code assumes logical pixels for all
transforms that are computed so we must set the viewport/window
transformations to account for the possibilty that widget
width in logical coordinates does not match the width in
devce pixels.