Modify processTubes
Code previously only went up to the grandparent level if the
parent of the supplied tube only had a single child. Broaden this
out to always go up to the grandparent level and the tube set
found isn't flat, fall back to considering the parent level. This
will help identify banks in the Bilby instrument

Also remove the code in calcBankRotation that returned an empty
Quat - didn't work for case where normal and z axis were 180 degrees
apart
Rotate bank normal to +ve or -ve z depending on whether bank is at
-ve or +ve z respectively