Spawn a process so that atexit will shutdown the Framework
The multiprocessing module supports different start methods for creating a
process. On unix the default is called fork. This method means the child
process will pick up environment variables from the parent process, such as
std::once_flag variables. If this has already been set on the parent process,
the relevant atexit code will not be executed on the child process. Therefore
when exit happens on the child process which is running MantidWorkbench,
the FrameworkManager and other services will not be shutdown. The solution
is to use the spawn start method which creates an entirely new python process
which will not inherit the std::once_flag variables.