Check for valid shortened file paths
Make use of the FileFinder singleton to find file paths. This means that
just giving the filename on its own should be enough to reach the file.
The algorithms underneath do this anyway, so this is mainly to make sure
that the validation on the GUI matches up with the valid paths for the
algortithms.

RE #34720