Keep the order of events when appending
If either this or the other EventList is empty, keep the order as
appropriate. This should save on later operations that try re-establish
the sort order. It will be most noticable when processing data in chunks
which will create empty EventLists from the part of the workspace that
is not currently being processed at the point of bringing everything
together again.